requirements.txt
aiolimiter==1.1.0
anyio==4.2.0
certifi==2023.11.17
dnspython==2.4.2
greenlet==3.0.1
h11==0.14.0
httpcore==1.0.2
httpx==0.26.0
idna==3.6
markdown-it-py==3.0.0
mdurl==0.1.2
motor==3.3.2
numpy==1.26.3
pandas==2.1.4
playwright==1.40.0
pyee==11.0.1
Pygments==2.17.2
pymongo==4.6.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
rich==13.7.0
selectolax==0.3.17
six==1.16.0
sniffio==1.3.0
typing_extensions==4.9.0
tzdata==2023.4
undetected-playwright==0.2.0
